+++ src/ert/_c_wrappers/enkf/config/enkf_config_node.py
@@ -6,7 +6,6 @@ from cwrap import BaseCClass
 from ecl.util.util import IntVector, StringList
 
 from ert._c_wrappers import ResPrototype
-from ert.parsing import ConfigValidationError
 from ert._c_wrappers.enkf.config_keys import ConfigKeys
 from ert._c_wrappers.enkf.enums import (
     EnkfTruncationType,
@@ -14,6 +13,7 @@ from ert._c_wrappers.enkf.enums import (
     FieldFileFormatType,
     LoadFailTypeEnum,
 )
+from ert.parsing import ConfigValidationError
 
 from .ext_param_config import ExtParamConfig
 from .field_config import FieldConfig
