+++ src/ert/_c_wrappers/enkf/enkf_obs.py
@@ -6,9 +6,9 @@ from ecl.util.util import CTime, StringList
 
 from ert import _clib
 from ert._c_wrappers import ResPrototype
-from ert.parsing import ConfigValidationError
 from ert._c_wrappers.enkf.enums import EnkfObservationImplementationType
 from ert._c_wrappers.enkf.observations import ObsVector
+from ert.parsing import ConfigValidationError
 
 if TYPE_CHECKING:
     from ecl.summary import EclSum
