+++ src/ert/_c_wrappers/job_queue/ext_job.py
@@ -9,9 +9,9 @@ from pathlib import Path
 from typing import Dict, List, Optional, Tuple
 
 from ert._c_wrappers.config import ConfigParser, ContentTypeEnum
-from ert.parsing import ConfigValidationError
 from ert._c_wrappers.util import SubstitutionList
 from ert._clib.job_kw import type_from_kw
+from ert.parsing import ConfigValidationError
 
 _SUBSTITUTED_AT_EXECUTION_TIME: List[str] = ["<ITER>", "<IENS>"]
