+++ tests/unit_tests/c_wrappers/res/enkf/test_analysis_config.py
@@ -1,7 +1,7 @@
 import pytest
 
-from ert.parsing import ConfigValidationError
 from ert._c_wrappers.enkf import AnalysisConfig, ConfigKeys
+from ert.parsing import ConfigValidationError
 
 
 @pytest.fixture
