+++ tests/unit_tests/c_wrappers/res/enkf/test_ensemble_config.py
@@ -5,9 +5,9 @@ import pytest
 from ecl.grid.ecl_grid import EclGrid
 from ecl.summary import EclSum
 
-from ert.parsing import ConfigWarning, ConfigValidationError
 from ert._c_wrappers.enkf import ConfigKeys, EnsembleConfig, ErtConfig
 from ert._c_wrappers.enkf.enums import ErtImplType, GenDataFileType
+from ert.parsing import ConfigValidationError, ConfigWarning
 
 
 def test_create():
