+++ tests/unit_tests/c_wrappers/res/enkf/test_ert_config.py
@@ -10,12 +10,12 @@ from unittest.mock import MagicMock, mock_open, patch
 import pytest
 from ecl.util.enums import RngAlgTypeEnum
 
-from ert.parsing import ConfigValidationError
 from ert._c_wrappers.enkf import AnalysisConfig, ConfigKeys, ErtConfig, HookRuntime
 from ert._c_wrappers.enkf._config_content_as_dict import parse_signature_job
 from ert._c_wrappers.enkf.ert_config import site_config_location
 from ert._c_wrappers.job_queue import QueueDriverEnum
 from ert._c_wrappers.sched import HistorySourceEnum
+from ert.parsing import ConfigValidationError
 
 config_defines = {
     "<USER>": "TEST_USER",
