+++ tests/unit_tests/c_wrappers/res/job_queue/test_ext_job.py
@@ -5,8 +5,8 @@ import stat
 import pytest
 
 from ert._c_wrappers.config import ContentTypeEnum
-from ert.parsing import ConfigValidationError
 from ert._c_wrappers.job_queue.ext_job import ExtJob
+from ert.parsing import ConfigValidationError
 
 
 @pytest.mark.usefixtures("use_tmpdir")
