+++ tests/unit_tests/c_wrappers/res/job_queue/test_function_ert_script.py
@@ -1,7 +1,7 @@
 import pytest
 
-from ert.parsing import ConfigValidationError
 from ert._c_wrappers.job_queue import WorkflowJob
+from ert.parsing import ConfigValidationError
 
 from .workflow_common import WorkflowCommon
